refactor(kernel): split contracts/ into schemas/ and events/ with typed domain events
Reorganizes kernel/ to match the kernel-and-shared-types standard:
- DTOs moved to kernel/schemas/ (one per file)
- Events moved to kernel/events/ (one per file)
- DomainEvent no longer has generic payload: dict — events use typed fields
- contracts/ converted to backward-compatible re-export shims
