+++ backend/src/hiresense/kernel/contracts/optimization.py
@@ -1,16 +1,6 @@
-from __future__ import annotations
+"""Backward-compatible re-export. Import from kernel.schemas."""
 
-from pydantic import BaseModel
+from hiresense.kernel.schemas.optimization_request_dto import OptimizationRequestDTO
+from hiresense.kernel.schemas.tex_diff_dto import TexDiffDTO
 
-
-class OptimizationRequestDTO(BaseModel):
-    match_id: str
-    job_id: str
-    cv_id: str
-
-
-class TexDiffDTO(BaseModel):
-    optimization_id: str
-    original_path: str
-    diff: str
-    modified_tex: str
+__all__ = ["OptimizationRequestDTO", "TexDiffDTO"]
