+++ backend/src/hiresense/kernel/events/__init__.py
@@ -1,0 +1,5 @@
+from hiresense.kernel.events.base import DomainEvent
+from hiresense.kernel.events.jobs_ingested import JobsIngestedEvent
+from hiresense.kernel.events.match_completed import MatchCompletedEvent
+
+__all__ = ["DomainEvent", "JobsIngestedEvent", "MatchCompletedEvent"]
