+++ backend/src/hiresense/kernel/events/base.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from datetime import datetime, timezone
-from typing import Any
 
 from pydantic import BaseModel, Field
 
@@ -9,4 +8,3 @@ from pydantic import BaseModel, Field
 class DomainEvent(BaseModel):
     event_type: str
     timestamp: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    payload: dict[str, Any] = Field(default_factory=dict)
