+++ backend/src/hiresense/kernel/schemas/__init__.py
@@ -1,0 +1,15 @@
+from hiresense.kernel.schemas.candidate_skills_dto import CandidateSkillsDTO
+from hiresense.kernel.schemas.cv_embedding_dto import CVEmbeddingDTO
+from hiresense.kernel.schemas.match_result_dto import MatchResultDTO
+from hiresense.kernel.schemas.normalized_job_dto import NormalizedJobDTO
+from hiresense.kernel.schemas.optimization_request_dto import OptimizationRequestDTO
+from hiresense.kernel.schemas.tex_diff_dto import TexDiffDTO
+
+__all__ = [
+    "CandidateSkillsDTO",
+    "CVEmbeddingDTO",
+    "MatchResultDTO",
+    "NormalizedJobDTO",
+    "OptimizationRequestDTO",
+    "TexDiffDTO",
+]
