+++ backend/src/hiresense/kernel/schemas/candidate_skills_dto.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from pydantic import BaseModel
+
+
+class CandidateSkillsDTO(BaseModel):
+    skills: list[str]
+    experience_summary: str
+    language: str
