+++ backend/src/hiresense/kernel/schemas/cv_embedding_dto.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from pydantic import BaseModel
+
+
+class CVEmbeddingDTO(BaseModel):
+    cv_id: str
+    embedding: list[float]
