+++ backend/src/hiresense/kernel/schemas/tex_diff_dto.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from pydantic import BaseModel
+
+
+class TexDiffDTO(BaseModel):
+    optimization_id: str
+    original_path: str
+    diff: str
+    modified_tex: str
